+++ examples/data-sources/auth0_organization_client/data-source.tf
@@ -1,0 +1,4 @@
+data "auth0_organization_client" "my_org_client" {
+  organization_id = "org_XXXXX"
+  client_id       = "clientXXXXX"
+}